third_party/WebKit/LayoutTests/virtual/serviceworker/http/tests/serviceworker/indexeddb-expected.txt

Verify that IndexedDB is functional in a ServiceWorker

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


messageChannel = new MessageChannel()
messageChannel.port1.onmessage = onMessageHandler

navigator.serviceWorker.register('indexeddb-worker.js')
serviceWorker.postMessage({port: messageChannel.port2}, [messageChannel.port2])
[ServiceWorker] Preparing the database in the service worker
[ServiceWorker] indexedDB.deleteDatabase('db')
[ServiceWorker] indexedDB.open('db')
[ServiceWorker] db.createObjectStore('store')
[ServiceWorker] tx = db.transaction('store', 'readwrite')
[ServiceWorker] store = tx.objectStore('store')
[ServiceWorker] store.put('value', 'key')

Verifying the database from the page

request = indexedDB.open('db')
db = request.result
tx = db.transaction('store')
store = tx.objectStore('store')
request = store.get('key')
PASS request.result is 'value'
PASS successfullyParsed is true

TEST COMPLETE
